Feat: added exam command
The goal is approximately the same as that of the "learn" command, but instead of choosing answers, the user must enter the whole word without prompts. The percentage of correct answers is also considered.

Fixed check functions. I did not consider that the message could be sent from another channel. Also, a bot can accidentally reply to its own message.

Signed-off-by: gutsy47 <bestcom5055@gmail.com>